[ feat ] 調整で細かい機能の追加
